# Version 1.0 Commit - 수정

+++ KaKaoProgram/Inc/ChattingRoom.h
@@ -22,11 +22,11 @@ public:
 	void SetChatDataList(std::list<ChattingData> datalist);
 	void SetChatData(ChattingData data);
 	void SetRoomName(const std::string name);
-	void SetRoomID(const INT64 roomID);
+	void SetRoomID(const int roomID);
 	void Read(Stream&  instream);
 	void Write(Stream& outStream);
 
-	inline INT64 GetRoomID()const { return m_RoomID; }
+	inline int GetRoomID()const { return m_RoomID; }
 	inline std::string GetRoomName()const { return m_RoomName; }
 	inline std::vector<std::string> GetJoinnedUsers()const { return m_JoinnedUsers; }
 	inline std::list<ChattingData> GetChattingDataList()const { return m_ChattingDataList; }
@@ -34,5 +34,5 @@ private:
 	std::vector<std::string> m_JoinnedUsers;
 	std::list<ChattingData> m_ChattingDataList;
 	std::string m_RoomName;
-	INT64 m_RoomID;
+	int m_RoomID;
 };
